+++ src/main/java/lotto/dto/LottoRequestDto.java
@@ -1,6 +1,6 @@
 package lotto.dto;
 
-import lotto.Lotto;
+import lotto.domain.Lotto;
 
 import java.util.List;
