+++ src/main/java/lotto/global/util/LottoGenerator.java
@@ -1,7 +1,7 @@
 package lotto.global.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 
 import java.util.ArrayList;
 import java.util.List;
